Return 404 on delete when not existing.

+++ src/main/java/de/sample/spring/customers/persistence/DatabaseCustomerSink.java
@@ -42,8 +42,12 @@ public class DatabaseCustomerSink implements CustomerSink {
     }
 
     @Override
-    public void delete(UUID id) {
-        // repo.existsById(id)
-        repo.deleteById(id);
+    public boolean delete(UUID id) {
+        if(repo.existsById(id)) {
+            repo.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
